requirements.txt
pandas
numpy
pdfplumber
sentence-transformers
faiss-cpu
langchain
gradio
PyMuPDF
openpyxl
matplotlib
python-dateutil